Front/fix

See merge request s04-webmobile2-sub3/s04p13d110!32

+++ frontend/src/views/photo/PhotoView.vue
@@ -126,8 +126,17 @@
           color="#DDA288"
           style="color:white; font-weight:bold;"
         >
-          게시글
-          삭제
+          <v-progress-linear
+            v-if="fab"
+            indeterminate
+            color="red"
+            style="width: 70px"
+          ></v-progress-linear>
+          <div
+            v-else
+          >
+            게시글 삭제
+          </div>
         </v-btn>
       </div>
     </div>
@@ -154,6 +163,7 @@ export default {
     FluxPreloader
   },
   data: () => ({
+    fab: false,
     vfOptions: {
       autoplay: false
     },
@@ -210,12 +220,17 @@ export default {
     },
     // 게시글 삭제
     deleteArticle: function () {
+      if (this.fab == true) {
+        this.fab = false
+        return
+      }
       // 이거 id로 바꾼 후, 다시 프로필의 id를 받아와서 보안성 높이기.
       if (this.$store.state.Auth.authToken.username != this.author) {
         console.log(this.$store.state.Auth)
         alert('인증되지 않은 사용자 입니다.')
       }
       else {
+        this.fab = true
         axios.delete (
           `${SERVER.BOARD_BASE_URL}delpost?id=${localStorage.getItem('articleId')}`,
           {
